Updated login route to match the api
Updated register route to match the api

Added an async logout function to hit the /api/auth/logout route on the api.
This is to let the api know that a user has logged out and that the token
that is being used to recognize them is no longer valid.